use double instead of int in factorial function

+++ src/interp.c
@@ -1291,11 +1291,11 @@ double eval(register struct ent * ent, register struct enode * e) {
     case WHITE:  return ((double) COLOR_WHITE);
     case FACT:
         {
-            int total = eval(ent, e->e.o.left);
+            double total = eval(ent, e->e.o.left);
             for (int i = eval(ent, e->e.o.left) - 1; i > 0; i--) {
                 total *= i;
-	    }
-	    return total > 0 ? total : 1;
+            }
+            return total > 0 ? total : 1;
         }
     default:    sc_error ("Illegal numeric expression");
                 exprerr = 1;
